HW12: rename fabric to factory
