+++ hw12_13_14_15_calendar/internal/storage/factory/factory.go
@@ -1,4 +1,4 @@
-package fabric
+package factory
 
 import (
 	"context"
